requirements.txt
torch
torchsde
torchvision
torchaudio
einops
transformers>=4.25.1
safetensors>=0.4.2
aiohttp
pyyaml
Pillow
scipy
tqdm
psutil

#non essential dependencies:
kornia>=0.7.1
spandrel
soundfile
marshmallow
runpod
boto3
cuid
insightface>=0.7.3
onnxruntime
onnxruntime-gpu
python-dotenv
uvicorn
fastapi
setuptools==69.5.1
opencv-python-headless>=4.10.0.82
cmake
dlib==19.24.2
